requirements.txt
beautifulsoup4==4.12.3
Bio==1.6.2
biopython==1.83
customtkinter==5.2.0
Pillow==10.2.0
pymolPy3==0.1.2
pymolPy3==0.1.2
pyperclip==1.8.2
Requests==2.31.0
tkinter_tooltip==3.0.0
Wikipedia_API==0.6.0
numpy>=1.25.2
matplotlib>=3.7.2
tkinter-tooltip==3.0.0
scipy>=1.12.0
pywaffle~=1.1.0
future~=0.18.3
PyYAML~=6.0.1